Remove some useless files.
We don't use third-party package configuration files, so I removed the
.gitkeep for that.

There's no need for the empty .gitignore in the database folder.

We use nginx for our server, and since the .htaccess is an Apache config
file it doesn't make sense to keep it.

bower.json is a third-party package dependency tool. It doesn't make sense
to put it in the public folder, and since it's not used for anything at
the moment, there is no real need to keep it.

favicon.ico was empty and the real one is in the public/img folder.

The databases are no longer used (if PR #34 is merged).

The maps are no longer used (if PR #33 is merged).

Signed-off-by: Austin Hellyer <hello@austinhellyer.me>